pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strategickhaos-document-refinery"
version = "0.1.0"
description = "Sovereign AI Document Intelligence Lab - Zero vendor lock-in AI legal/document review layer"
readme = "document_refinery/README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Strategickhaos DAO LLC", email = "node137@strategickhaos.dev"}
]
keywords = ["ai", "document-review", "legal-tech", "red-team", "blue-team", "purple-team", "sovereignty"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "tiktoken>=0.5.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.282",
    "mypy>=1.5.0",
]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
local = [
    "transformers>=4.30.0",
    "torch>=2.0.0",
]

[project.scripts]
sk-refinery = "document_refinery.cli.main:cli"

[tool.setuptools.packages.find]
include = ["document_refinery*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
